Move scener_with_affiliations template tag to releaser_tags
