+++ demoscene/templates/sceners/history.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% load demoscene_tags compress %}
+{% load releaser_tags compress %}
 
 
 {% block html_title %}{{ scener.name_with_affiliations }} edit history - Demozoo{% endblock %}
