+++ demoscene/templatetags/demoscene_tags.py
@@ -3,29 +3,12 @@ from django.template.defaultfilters import date as date_format
 from django.utils.html import format_html
 
 from bbs.models import BBS
-from demoscene.models import Edit, Nick, Releaser
+from demoscene.models import Edit, Releaser
 from parties.models import Party
 from productions.models import Production
 
 
 register = template.Library()
-
-
-@register.inclusion_tag("tags/scener_with_affiliations.html")
-def scener_with_affiliations(releaser_or_nick):
-    if isinstance(releaser_or_nick, Nick):
-        releaser = releaser_or_nick.releaser
-    else:  # assume a Releaser
-        releaser = releaser_or_nick
-    groups = releaser.current_groups()
-
-    return {
-        "name": releaser_or_nick.name,
-        "releaser": releaser,
-        "groups": groups,
-        # use abbreviations for groups if total length of names is 20 or more chars
-        "abbreviate_groups": (sum([len(group.name) for group in groups]) >= 20),
-    }
 
 
 @register.inclusion_tag("tags/icon.html")
