+++ homepage/templates/homepage/activity/latest_activity.html
@@ -1,5 +1,5 @@
 {% extends 'base.html' %}
-{% load demoscene_tags compress %}
+{% load demoscene_tags releaser_tags %}
 
 {% block base_main %}
     <header>
