+++ maintenance/templates/maintenance/matching_real_names.html
@@ -1,5 +1,5 @@
 {% extends "maintenance/base.html" %}
-{% load demoscene_tags %}
+{% load demoscene_tags releaser_tags %}
 
 {% block html_title %}{{ title }}{% endblock %}
