+++ search/templates/search/results/scener.html
@@ -1,4 +1,4 @@
-{% load demoscene_tags %}
+{% load demoscene_tags releaser_tags %}
 <div class="result">
     <div class="media media--thumbnail">
         {% icon 'scener' %}
